+++ topiary/__init__.py
@@ -1,4 +1,8 @@
-from .predictor import TopiaryPredictor, from_predictions
+from .predictor import (
+    TopiaryPredictor,
+    fragment_from_effect,
+    from_predictions,
+)
 from .ranking import (
     self_nearest,
     Affinity,
@@ -22,6 +26,7 @@ from .ranking import (
     mhc_dependence,
     resolve_default_methods,
     describe_default_versions,
+    is_named_version,
     resolve_default_versions,
     validate_default_methods,
     validate_default_versions,
@@ -96,11 +101,12 @@ from .result import (
 )
 from .wide import detect_form, from_wide, to_wide
 
-__version__ = "5.34.0"
+__version__ = "5.35.0"
 
 __all__ = [
     "TopiaryPredictor",
     "from_predictions",
+    "fragment_from_effect",
     "CachedPredictor",
     "mhcflurry_composite_version",
     "SelfProteome",
@@ -125,6 +131,7 @@ __all__ = [
     "mhc_dependence",
     "resolve_default_methods",
     "describe_default_versions",
+    "is_named_version",
     "resolve_default_versions",
     "validate_default_methods",
     "validate_default_versions",